% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzurePipe.R
\name{AzurePipe}
\alias{AzurePipe}
\title{Azure Blob Storage Pipeline R6 Class}
\value{
A list of `blob_container` objects
}
\description{
Extends `PipelineBase` to provide methods for interacting with Azure
Blob Storage. Inherits parameter handling and logging.
}
\section{Super class}{
\code{\link[constRuct:PipelineBase]{constRuct::PipelineBase}} -> \code{AzurePipe}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{get_containers}}{List of available container objects for the active endpoint. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AzurePipe-new}{\code{AzurePipe$new()}}
\item \href{#method-AzurePipe-list_blobs}{\code{AzurePipe$list_blobs()}}
\item \href{#method-AzurePipe-get_blob_metadata}{\code{AzurePipe$get_blob_metadata()}}
\item \href{#method-AzurePipe-upload_log_file}{\code{AzurePipe$upload_log_file()}}
\item \href{#method-AzurePipe-clone}{\code{AzurePipe$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="constRuct" data-topic="PipelineBase" data-id="log_read"><a href='../../constRuct/html/PipelineBase.html#method-PipelineBase-log_read'><code>constRuct::PipelineBase$log_read()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzurePipe-new"></a>}}
\if{latex}{\out{\hypertarget{method-AzurePipe-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzurePipe$new(pipeline_name, pipeline_id, division = NULL, program = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pipeline_name}}{Name of this pipeline.}

\item{\code{pipeline_id}}{Identifier for this pipeline}

\item{\code{division}}{The division this data belongs to.}

\item{\code{program}}{The program subset this data is for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `AzurePipe` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzurePipe-list_blobs"></a>}}
\if{latex}{\out{\hypertarget{method-AzurePipe-list_blobs}{}}}
\subsection{Method \code{list_blobs()}}{
Lists blobs in a specified Azure container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzurePipe$list_blobs(container_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{container_name}}{The name of the container (character).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame of blob properties from `AzureStor::list_blobs`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzurePipe-get_blob_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-AzurePipe-get_blob_metadata}{}}}
\subsection{Method \code{get_blob_metadata()}}{
Gets the user-defined metadata for a specific blob.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzurePipe$get_blob_metadata(container_name, blob_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{container_name}}{The name of the container where the blob resides.}

\item{\code{blob_name}}{The name of the blob.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of metadata key-value pairs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzurePipe-upload_log_file"></a>}}
\if{latex}{\out{\hypertarget{method-AzurePipe-upload_log_file}{}}}
\subsection{Method \code{upload_log_file()}}{
Uploads text content as a log file to a specified container.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzurePipe$upload_log_file(file_content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_content}}{A character vector containing the text to upload.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzurePipe-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AzurePipe-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzurePipe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
