% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.comparison.R
\name{calculate.layer.importance}
\alias{calculate.layer.importance}
\title{Calculate layer importance}
\usage{
calculate.layer.importance(conStruct.results, data.block, layer.order = NULL)
}
\arguments{
\item{conStruct.results}{The list output by a 
\code{conStruct} run for a given MCMC chain.}

\item{data.block}{A \code{data.block} list saved during a 
\code{conStruct} run.}

\item{layer.order}{An optional \code{vector} giving the
order in which the clusters of \code{conStruct.results} are 
    read.}
}
\value{
This function returns a \code{vector} giving the 
			relative contributions (importance) of the clusters 
			in the analysis.
}
\description{
\code{calculate.layer.importance}
}
\details{
This function takes the results of a \code{conStruct} 
analysis and calculates the relative contributions of 
each cluster to total covariance.

This function calculates the contribution of each cluster to
		total covariance by multiplying the within-cluster covariance 
		in a given cluster by the admixture proportions samples draw 
		from that cluster. The relative contribution of that cluster, 
		which can be interpreted as its importance, is this absolute 
		contribution divided by the sum of those of all other clusters. 
			A cluster can have a large importance if many samples draw 
		large amounts of admixture from it, or if it has a very large 
		within-cluster covariance parameter (phi), or some combination 
		of the two. Cluster importance can be useful for evaluating 
		an appropriate level of model complexity for the data (e.g., 
		choosing a value of \code{K} or comparing the spatial and 
		nonspatial models).
}
