% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_env_vars}
\alias{check_env_vars}
\title{Check a named list of environment variables}
\usage{
check_env_vars(env_vars)
}
\arguments{
\item{env_vars}{A named list where names are the environment variable names and
values are their retrieved values.}
}
\value{
Returns `invisible(TRUE)` if all variables are valid.
}
\description{
This function iterates through a named list of values, checks if any are
  invalid (NULL, NA, or an empty string), and throws a descriptive error if any
  are found.
}
