% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conStruct.data}
\alias{conStruct.data}
\title{Example dataset used in a \code{conStruct} analysis}
\format{A list with two elements:
\describe{
	\item{allele.frequencies}{a matrix with one row for each of 
		the 16 samples and one column for each of 10,000 loci, 
		giving the frequency of the counted allele at each locus 
		in each sample}
	\item{coords}{a matrix with one row for each of the 16 samples, 
		in the same order as that of the allele frequency matrix, 
		and two columns, the first giving the x-coordinate 
		(or longitude), the second giving the y-coordinate (or latitude)}
}}
\usage{
conStruct.data
}
\description{
A simulated dataset containing the allele frequency 
and sampling coordinate data necessary to run a 
\code{conStruct} analysis.
}
\keyword{datasets}
