% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conStruct.results}
\alias{conStruct.results}
\title{Example model output from a \code{conStruct} analysis}
\format{a list with one entry for each chain run 
	(specified with \code{n.chains}). The entry for each chain is named 
	"chain_X" for the Xth chain.  The components of the entries for each 
	are detailed below: 
	\itemize{
		\item \code{posterior} gives parameter estimates over the posterior 
				distribution of the MCMC
			\itemize{
				\item \code{n.iter} number of MCMC iterations retained for 
						analysis (half of the \code{n.iter} argument 
						specified in the function call)
				\item \code{lpd} vector of log posterior density over the retained 
						MCMC iterations.
				\item \code{nuggets} matrix of estimated nugget parameters with 
						one row per MCMC iteration and one column per sample.
				\item \code{par.cov} array of estimated parametric covariance matrices, 
						for which the first dimension is the number of MCMC iterations.
				\item \code{gamma} vector of estimated gamma parameter.
				\item \code{layer.params} list summarizing estimates of layer-specific 
						parameters. There is one entry for each layer specified, and the 
						entry for the kth layer is named "Layer_k".
					\itemize{
						\item \code{alpha0} vector of estimated alpha0 parameter in the 
								kth layer.
						\item \code{alphaD} vector of estimated alphaD parameter in the 
								kth layer.
						\item \code{alpha2} vector of estimated alpha2 parameter in the 
								kth layer.
						\item \code{mu} vector of estimated mu parameter in the 
								kth layer.
						\item \code{layer.cov} vector of estimated layer-specific 
								covariance parameter in the kth layer.
					}
				\item \code{admix.proportions} array of estimated admixture proportions.
						The first dimension is the number of MCMC iterations, 
						the second is the number of samples, 
							and the third is the number of layers.
			}
	\item \code{MAP} gives point estimates of the parameters listed in the \code{posterior}
						list described above. Values are indexed at the MCMC iteration 
						with the greatest posterior probability.
			\itemize{
				\item \code{index.iter} the iteration of the MCMC with the highest 
						posterior probability, which is used to index all parameters 
						included in the \code{MAP} list
				\item \code{lpd} the greatest value of the posterior probability
				\item \code{nuggets} point estimate of nugget parameters
				\item \code{par.cov} point estimate of parametric covariance
				\item \code{gamma} point estimate of gamma parameter
				\item \code{layer.params} point estimates of all layer-specific parameters 
				\item \code{admix.proportions} point estimates of admixture proportions.
			}
	}}
\usage{
conStruct.results
}
\description{
An example \code{conStruct.results} object generated by 
a \code{conStruct} analysis.
}
\keyword{datasets}
