% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.system_snapshot}
\alias{.system_snapshot}
\title{Snapshot the system information}
\usage{
.system_snapshot()
}
\value{
A list with `config`, `sysname`, `user`, `node`
}
\description{
Snapshot the system information
}
