% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.output.R
\name{make.structure.plot}
\alias{make.structure.plot}
\title{Make STRUCTURE output plot}
\usage{
make.structure.plot(admix.proportions, mar = c(2, 4, 2, 2),
  sample.order = NULL, layer.order = NULL, sample.names = NULL,
  sort.by = NULL, layer.colors = NULL)
}
\arguments{
\item{admix.proportions}{A \code{matrix} of admixture proportions, 
with one row per sample and one column per layer.}

\item{mar}{A \code{vector} of plotting margins passed to \code{par}.
Default is \code{c(2,4,2,2)}, which tends to look good.}

\item{sample.order}{A \code{vector} giving the order in which sample 
admixture proportions are to be plotted, left to right.  If 
\code{NULL}, samples are plotted in the order they occur in 
\code{admix.proportions}.}

\item{layer.order}{A \code{vector} giving the order in which layers 
are plotted, bottom to top. If \code{NULL}, layers are plotted 
in the order they occur in \code{admix.proportions}.}

\item{sample.names}{Vector of names to be plotted under each sample's 
admixture proportion bar plot. The index of a sample's name 
should be the same as the index of the sample's row in 
\code{admix.proportions}. If \code{NULL}, no names 
are printed.}

\item{sort.by}{An \code{integer} giving the column index of the \code{admix.proportions} 
matrix to be used in determining sample plotting order.  If specified, 
samples are plotted from left to right in increasing order of their 
membership in that layer.  If \code{NULL}, samples are plotted 
in the order they occur in \code{admix.proportions}.}

\item{layer.colors}{A \code{vector} of colors to be used in plotting 
results for different layers. Users must specify one 
color per layer.  If \code{NULL}, the plot will use 
a pre-specified vector of colors.}
}
\value{
This function has only invisible return values.
}
\description{
\code{make.structure.plot} makes a STRUCTURE-style plot from the output from a 
	conStruct analysis.
}
\details{
This function takes the output from a conStruct analysis and 
makes a STRUCTURE-style plot, where each sample 
is represented as a stacked bar plot, and the length of the 
bar plot segments of each color represent that sample's 
admixture proportion in that layer.
}

