% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLikelihood.R
\name{maxLikelihoodOld}
\alias{maxLikelihoodOld}
\title{max.likelihood}
\usage{
maxLikelihoodOld(data, max.alleles, resolution)
}
\arguments{
\item{data}{the input file}

\item{max.alleles}{max.alleles places an uppermost limit on the number of alleles considered}

\item{resolution}{resolution is the resolution of the F parameter}
}
\description{
max.likelihood
}
\examples{
inFile = system.file("extdata", "infile.txt", package = "ConStruct", mustWork = TRUE)
maxLikelihoodOld(data = inFile, max.alleles = 1000, resolution = 100)
}
