% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineBase.R
\name{PipelineBase}
\alias{PipelineBase}
\title{ETL Pipeline R6 Class}
\value{
Returns the path to the log file.
}
\description{
A base R6 class for creating and managing stateful ETL pipelines. It handles
parameter validation, dataset storage, and logging.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pipeline_name}}{(`character()`)\cr
A title case name for the pipeline e.g. 'Pipeline Name'}

\item{\code{pipeline_id}}{(`character()`)\cr
Unique Identifier for this pipeline e.g. 'pipeline-id'}

\item{\code{metadata}}{(`list()`)\cr
A list to hold arbitrary user-defined pipeline attributes or identifiers.}

\item{\code{datasets}}{(`list()`)\cr
A list to store datasets (e.g., data frames) used or created.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipelineBase-new}{\code{PipelineBase$new()}}
\item \href{#method-PipelineBase-set_meta}{\code{PipelineBase$set_meta()}}
\item \href{#method-PipelineBase-get_meta}{\code{PipelineBase$get_meta()}}
\item \href{#method-PipelineBase-add_dataset}{\code{PipelineBase$add_dataset()}}
\item \href{#method-PipelineBase-get_dataset}{\code{PipelineBase$get_dataset()}}
\item \href{#method-PipelineBase-add_log}{\code{PipelineBase$add_log()}}
\item \href{#method-PipelineBase-get_logs}{\code{PipelineBase$get_logs()}}
\item \href{#method-PipelineBase-end_pipeline}{\code{PipelineBase$end_pipeline()}}
\item \href{#method-PipelineBase-print}{\code{PipelineBase$print()}}
\item \href{#method-PipelineBase-clone}{\code{PipelineBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$new(pipeline_name, pipeline_id, metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pipeline_name}}{(`character()`)\cr
Name of this pipeline.}

\item{\code{pipeline_id}}{(`character()`)\cr
Identifier for this pipeline}

\item{\code{metadata}}{(`list()`)\cr
An optional named list of attributes to add}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PipelineBase` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-set_meta"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-set_meta}{}}}
\subsection{Method \code{set_meta()}}{
Add or update a single piece of metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$set_meta(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A character string for the metadata name (the key).}

\item{\code{value}}{The value to be stored. Can be any R object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-get_meta"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-get_meta}{}}}
\subsection{Method \code{get_meta()}}{
Retrieve a metadata value for a given key.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$get_meta(key, default = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The name of the metadata attribute to retrieve.}

\item{\code{default}}{The value to return if the key is not found.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The metadata value, or the default value if not found.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-add_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-add_dataset}{}}}
\subsection{Method \code{add_dataset()}}{
Add a dataset to the pipeline's internal storage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$add_dataset(name, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name to assign to the dataset (character).}

\item{\code{data}}{The dataset object to store (e.g., a data frame).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-get_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-get_dataset}{}}}
\subsection{Method \code{get_dataset()}}{
Retrieve a dataset from the pipeline's storage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$get_dataset(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the dataset to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The dataset object associated with the given name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-add_log"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-add_log}{}}}
\subsection{Method \code{add_log()}}{
Add a timestamped message to the pipeline's log.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$add_log(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{The message string to add to the log.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-get_logs"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-get_logs}{}}}
\subsection{Method \code{get_logs()}}{
Retrieve all log messages.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$get_logs(formatted = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formatted}}{Logical. Whether to format logs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector of all log entries.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-end_pipeline"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-end_pipeline}{}}}
\subsection{Method \code{end_pipeline()}}{
Ends the pipeline run and writes the log to a local file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$end_pipeline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-print"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-print}{}}}
\subsection{Method \code{print()}}{
Overrides default print method. Prints the Pipeline Name and Metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Further arguments passed to or from other methods. (Currently not used).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
