% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.model.fit.R
\name{print.conStruct.results}
\alias{print.conStruct.results}
\title{An S3 print method for class conStruct.results}
\usage{
\method{print}{conStruct.results}(x, ...)
}
\arguments{
\item{x}{an object of class \code{conStruct.results}}

\item{...}{further options to be passed to \code{print}}
}
\value{
prints a top-level summary of the conStruct.results, returns nothing
}
\description{
An S3 print method for class conStruct.results
}
