% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.conStruct.R
\name{print.data.block}
\alias{print.data.block}
\title{An S3 print method for class data.block}
\usage{
\method{print}{data.block}(x, ...)
}
\arguments{
\item{x}{an object of class \code{data.block}}

\item{...}{further options to be passed to \code{print}}
}
\value{
prints a top-level summary of the data.block, returns nothing
}
\description{
An S3 print method for class data.block
}
