% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.conStruct.R
\name{print.freq.data}
\alias{print.freq.data}
\title{An S3 print method for class freq.data}
\usage{
\method{print}{freq.data}(x, ...)
}
\arguments{
\item{x}{an object of class \code{freq.data}}

\item{...}{further options to be passed to \code{print}}
}
\value{
prints a top-level summary of the freq.data, returns nothing
}
\description{
An S3 print method for class freq.data
}
