% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.model.fit.R
\name{print.layer.params}
\alias{print.layer.params}
\title{An S3 print method for class layer.params}
\usage{
\method{print}{layer.params}(x, ...)
}
\arguments{
\item{x}{an object of class \code{layer.params}}

\item{...}{further options to be passed to \code{print}}
}
\value{
prints a top-level summary of the layer.params, returns nothing
}
\description{
An S3 print method for class layer.params
}
