% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate a population with desired levels of substructure and inbreeding}
\usage{
simulate(N, num.loc, fst, r.actual, c, r.consider, max.alleles,
  f.resolution, c.resolution, iteration)
}
\arguments{
\item{N}{the total sample size}

\item{num.loc}{the number of loci}

\item{fst}{the value of Fst that is to be simulated between two populations}

\item{r.actual}{the inbreeding coefficient of the inbred individuals}

\item{c}{the proportion of the population inbred to degree r.actual}

\item{r.consider}{the value of the inbreeding coefficient being considered for the}

\item{max.alleles}{places an uppermost limit on the number of alleles considered}

\item{f.resolution}{the resolution on the Fst parameter}

\item{c.resolution}{the resolution on the c parameter}

\item{iteration}{the number of iterations of the simulation run through in order}
}
\description{
Simulate a population with desired levels of substructure and inbreeding
}
\examples{
simulate(200, 12, 0.05, 0.05, 0.5, 0.05, 100, 100, 100, 10000)
}
