% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct}
\alias{construct}
\title{construct}
\usage{
construct(data, max.alleles, f.resolution, c.resolution, r)
}
\arguments{
\item{data}{the input file}

\item{max.alleles}{places an uppermost limit on the number of alleles considered}

\item{f.resolution}{the resolution of the Fst parameter}

\item{r}{is the value of the inbreeding coefficient being considered for the analysis of the dataset}

\item{the}{resolution on the c parameter}
}
\description{
construct
}
\examples{
construct(data = "infile.txt", max.alleles = 1000, f.resolution = 100, c.resolution = 100, r = 0.0625)
}
