% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLikelihood.R
\name{maxLikelihood}
\alias{maxLikelihood}
\title{maxLikelihood}
\usage{
maxLikelihood(data, max.alleles, resolution)
}
\arguments{
\item{data}{the input file}

\item{max.alleles}{max.alleles places an uppermost limit on the number of alleles considered}

\item{resolution}{resolution is the resolution of the F parameter}
}
\description{
maxLikelihood
}
\examples{
maxLikelihood(data = "infile.txt", max.alleles = 1000, resolution = 100)
}
