% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipelineBase.R
\name{PipelineBase}
\alias{PipelineBase}
\title{ETL Pipeline R6 Class}
\description{
A base R6 class for creating and managing stateful ETL pipelines. It handles
parameter validation, dataset storage, and logging.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Proper name of the pipeline}

\item{\code{id}}{Logging identifier for this pipeline.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PipelineBase-new}{\code{PipelineBase$new()}}
\item \href{#method-PipelineBase-log_read}{\code{PipelineBase$log_read()}}
\item \href{#method-PipelineBase-clone}{\code{PipelineBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$new(name, id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of this pipeline.}

\item{\code{id}}{Identifier for this pipeline, used in logging.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `PipelineBase` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-log_read"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-log_read}{}}}
\subsection{Method \code{log_read()}}{
Retrieves pipeline logs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$log_read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with each log entry
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PipelineBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PipelineBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PipelineBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
