% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read allele data from a file}
\usage{
readData(fileName, locusNames = FALSE, delim = "[,[:space:]]+",
  missing = NA)
}
\arguments{
\item{fileName}{}
}
\description{
Read allele data from a file
}
\examples{
inFile = system.file("extdata", "infile.txt", package = "ConStruct", mustWork = TRUE)
readData(inFile, missing = 0)
}
